+++ src/arguments.h
@@ -33,7 +33,6 @@ enum SHORT_ENUM Action {
     ACTION_RESUME,
     ACTION_STOP,
     ACTION_DUMP,
-    ACTION_CHECK,
     ACTION_STATUS,
     ACTION_METRICS,
     ACTION_LIST,
@@ -61,7 +60,6 @@ enum SHORT_ENUM CStack {
     CSTACK_NO,       // do not collect native frames
     CSTACK_FP,       // walk stack using Frame Pointer links
     CSTACK_DWARF,    // use DWARF unwinding info from .eh_frame section
-    CSTACK_LBR,      // Last Branch Record hardware capability
     CSTACK_VM        // unwind using HotSpot VMStructs
 };
 
@@ -107,25 +105,14 @@ enum EventMask {
 constexpr int EVENT_MASK_SIZE = 7;
 
 struct StackWalkFeatures {
-    // Deprecated stack recovery techniques used to workaround AsyncGetCallTrace flaws
-    unsigned short unknown_java  : 1;
-    unsigned short unwind_stub   : 1;
-    unsigned short unwind_comp   : 1;
-    unsigned short unwind_native : 1;
-    unsigned short java_anchor   : 1;
-    unsigned short gc_traces     : 1;
-
-    // Common features
     unsigned short stats         : 1;  // collect stack walking duration statistics
-
-    // Additional HotSpot-specific features
     unsigned short jnienv        : 1;  // verify JNIEnv* obtained using VMStructs
-    unsigned short probe_sp      : 1;  // when AsyncGetCallTrace fails, adjust SP and retry
+    unsigned short agct          : 1;  // force usage of AsyncGetCallTrace instead of VMStructs
     unsigned short mixed         : 1;  // mixed stack traces with Java and native frames interleaved
     unsigned short vtable_target : 1;  // show receiver classes of vtable/itable stubs
     unsigned short comp_task     : 1;  // display current compilation task for JIT threads
     unsigned short pc_addr       : 1;  // record exact PC address for each sample
-    unsigned short _padding      : 3;  // pad structure to 16 bits
+    unsigned short _padding      : 9;  // pad structure to 16 bits
 };
 
 
@@ -177,6 +164,7 @@ class Arguments {
     std::vector<const char*> _trace;
     int _timeout;
     int _loop;
+    size_t _mem_limit;
     long _interval;
     long _alloc;
     long _nativemem;
@@ -186,6 +174,7 @@ class Arguments {
     long _proc;
     bool _all;
     int _jstackdepth;
+    int _truncated_stack_depth;
     int _signal;
     const char* _file;
     const char* _log;
@@ -201,6 +190,7 @@ class Arguments {
     bool _threads;
     bool _sched;
     bool _record_cpu;
+    bool _tlab;
     bool _live;
     bool _nofree;
     bool _nobatch;
@@ -238,6 +228,7 @@ class Arguments {
         _trace(),
         _timeout(0),
         _loop(0),
+        _mem_limit(0),
         _interval(0),
         _alloc(-1),
         _nativemem(-1),
@@ -247,6 +238,7 @@ class Arguments {
         _proc(-1),
         _all(false),
         _jstackdepth(DEFAULT_JSTACKDEPTH),
+        _truncated_stack_depth(DEFAULT_JSTACKDEPTH),
         _signal(0),
         _file(NULL),
         _log(NULL),
@@ -262,6 +254,7 @@ class Arguments {
         _threads(false),
         _sched(false),
         _record_cpu(false),
+        _tlab(false),
         _live(false),
         _nofree(false),
         _nobatch(false),
@@ -271,7 +264,7 @@ class Arguments {
         _fdtransfer_path(NULL),
         _target_cpu(-1),
         _style(0),
-        _features{1, 1, 1, 1, 1, 1},
+        _features{},
         _cstack(CSTACK_DEFAULT),
         _clock(CLK_DEFAULT),
         _output(OUTPUT_NONE),
@@ -302,7 +295,7 @@ class Arguments {
 
     bool hasOutputFile() const {
         return _file != NULL &&
-            (_action == ACTION_STOP || _action == ACTION_DUMP ? _output != OUTPUT_JFR : _action >= ACTION_CHECK);
+            (_action == ACTION_STOP || _action == ACTION_DUMP ? _output != OUTPUT_JFR : _action >= ACTION_STATUS);
     }
 
     bool hasOption(JfrOption option) const {