+++ src/main/main.cpp
@@ -54,6 +54,7 @@ static const char USAGE_STRING[] =
     "  -g, --sig           print method signatures\n"
     "  -a, --ann           annotate Java methods\n"
     "  -l, --lib           prepend library names\n"
+    "  --dot               dotted class names\n"
     "  -o fmt              output format: flat|traces|collapsed|flamegraph|tree|jfr|otlp\n"
     "  -I include          output only stack traces containing the specified pattern\n"
     "  -X exclude          exclude stack traces with the specified pattern\n"
@@ -69,6 +70,7 @@ static const char USAGE_STRING[] =
     "\n"
     "  --loop time         run profiler in a loop\n"
     "  --alloc bytes       allocation profiling interval in bytes\n"
+    "  --tlab              use TLAB events for allocation profiling\n"
     "  --live              build allocation profile from live objects only\n"
     "  --nativemem bytes   native allocation profiling interval in bytes\n"
     "  --nofree            do not collect free calls in native allocation profiling\n"
@@ -76,19 +78,21 @@ static const char USAGE_STRING[] =
     "  --lock time         lock profiling threshold in nanoseconds\n"
     "  --nativelock time   pthread mutex/rwlock profiling threshold in nanoseconds\n"
     "  --wall interval     wall clock profiling interval\n"
+    "  --nobatch           legacy wall clock sampling without batch events\n"
     "  --proc interval     process sampling interval (default: 30s)\n"
     "  --all               shorthand for enabling cpu, wall, alloc, live,\n"
     "                      nativemem and lock profiling simultaneously\n"
     "  --total             accumulate the total value (time, bytes, etc.)\n"
     "  --all-user          only include user-mode events\n"
     "  --sched             group threads by scheduling policy\n"
-    "  --cstack mode       how to traverse C stack: fp|dwarf|lbr|vm|no\n"
+    "  --cstack mode       how to traverse C stack: fp|dwarf|vm|no\n"
     "  --signal num        use alternative signal for cpu or wall clock profiling\n"
     "  --clock source      clock source for JFR timestamps: tsc|monotonic\n"
     "  --begin function    begin profiling when function is executed\n"
     "  --end function      end profiling when function is executed\n"
     "  --ttsp              only time-to-safepoint profiling \n"
     "  --nostop            do not stop profiling outside --begin/--end window\n"
+    "  --memlimit bytes    limit size of the stack trace storage\n"
     "  --jfropts opts      JFR recording options: mem\n"
     "  --jfrsync config    synchronize profiler with JFR recording\n"
     "  --libpath path      full path to libasyncProfiler.so in the container\n"
@@ -412,7 +416,7 @@ int main(int argc, const char** argv) {
     while (args.count() > 0 && !(jattach_action && pid)) {
         String arg = args.next();
 
-        if (arg == "start" || arg == "resume" || arg == "stop" || arg == "dump" || arg == "check" ||
+        if (arg == "start" || arg == "resume" || arg == "stop" || arg == "dump" ||
             arg == "status" || arg == "metrics" || arg == "list" || arg == "collect") {
             action = arg;
 
@@ -500,30 +504,21 @@ int main(int argc, const char** argv) {
         } else if (arg == "--width" || arg == "--height" || arg == "--minwidth") {
             format << "," << (arg.str() + 2) << "=" << args.next();
 
-        } else if (arg == "--reverse" || arg == "--inverted" || arg == "--samples" || arg == "--total" ||
-                   arg == "--sched" || arg == "--live" || arg == "--nofree" || arg == "--record-cpu") {
+        } else if (arg == "--reverse" || arg == "--inverted" || arg == "--samples" || arg == "--total") {
             format << "," << (arg.str() + 2);
 
         } else if (arg == "--alloc" || arg == "--nativemem" || arg == "--nativelock" || arg == "--lock" ||
                    arg == "--wall" || arg == "--trace" || arg == "--chunksize" || arg == "--chunktime" ||
                    arg == "--cstack" || arg == "--signal" || arg == "--clock" || arg == "--begin" || arg == "--end" ||
-                   arg == "--target-cpu" || arg == "--proc") {
+                   arg == "--target-cpu" || arg == "--proc" || arg == "--memlimit") {
             params << "," << (arg.str() + 2) << "=" << args.next();
 
-        } else if (arg == "--ttsp") {
-            params << ",ttsp";
-
-        } else if (arg == "--nostop") {
-            params << ",nostop";
-
-        } else if (arg == "--all") {
-            params << ",all";
+        } else if (arg == "--all" || arg == "--live" || arg == "--nobatch" || arg == "--nofree" || arg == "--nostop" ||
+                   arg == "--record-cpu" || arg == "--sched" || arg == "--tlab" || arg == "--ttsp") {
+            params << "," << (arg.str() + 2);
 
         } else if (arg == "--all-user") {
             params << ",alluser";
-
-        } else if (arg == "--safe-mode") {
-            params << ",safemode=" << args.next();
 
         } else if (arg == "--jfrsync" || arg == "--jfropts") {
             params << "," << (arg.str() + 2) << "=" << args.next();