+++ src/stackFrame_arm.cpp
@@ -6,7 +6,6 @@
 #if defined(__arm__) || defined(__thumb__)
 
 #include <errno.h>
-#include <string.h>
 #include "stackFrame.h"
 #include "vmStructs.h"
 
@@ -68,37 +67,14 @@ void StackFrame::ret() {
 bool StackFrame::unwindStub(instruction_t* entry, const char* name, uintptr_t& pc, uintptr_t& sp, uintptr_t& fp) {
     instruction_t* ip = (instruction_t*)pc;
     if (ip == entry || *ip == 0xe12fff1e
-        || strncmp(name, "itable", 6) == 0
-        || strncmp(name, "vtable", 6) == 0
-        || strcmp(name, "InlineCacheBuffer") == 0)
+        || startsWith(name, "itable")
+        || startsWith(name, "vtable")
+        || streq(name, "InlineCacheBuffer"))
     {
         pc = link();
         return true;
     }
     return false;
-}
-
-bool StackFrame::unwindCompiled(NMethod* nm, uintptr_t& pc, uintptr_t& sp, uintptr_t& fp) {
-    instruction_t* ip = (instruction_t*)pc;
-    instruction_t* entry = (instruction_t*)nm->entry();
-    if (ip > entry && ip <= entry + 4 && (*ip & 0xffffff00) == 0xe24dd000) {
-        //    push  {r11, lr}
-        //    mov   r11, sp (optional)
-        // -> sub   sp, sp, #offs
-        fp = ((uintptr_t*)sp)[0];
-        pc = ((uintptr_t*)sp)[1];
-        sp += 8;
-        return true;
-    } else if (*ip == 0xe8bd4800) {
-        //    add   sp, sp, #offs
-        // -> pop   {r11, lr}
-        fp = ((uintptr_t*)sp)[0];
-        pc = ((uintptr_t*)sp)[1];
-        sp += 8;
-        return true;
-    }
-    pc = link();
-    return true;
 }
 
 bool StackFrame::unwindPrologue(NMethod* nm, uintptr_t& pc, uintptr_t& sp, uintptr_t& fp) {
@@ -123,10 +99,6 @@ bool StackFrame::unwindAtomicStub(const void*& pc) {
 
 void StackFrame::adjustSP(const void* entry, const void* pc, uintptr_t& sp) {
     // Not needed
-}
-
-bool StackFrame::skipFaultInstruction() {
-    return false;
 }
 
 bool StackFrame::checkInterruptedSyscall() {